refactor(messaging): move workflow ownership to desktop
