+++ apps/desktop/src/main/messaging/core/deterministic-interaction-mapper.ts
