+++ apps/desktop/src/main/messaging/core/interaction-mapper.ts
