+++ apps/desktop/src/main/messaging/core/messaging-adapter.ts
