+++ apps/desktop/src/main/messaging/core/messaging-approval-renderer.ts
