+++ apps/desktop/src/main/messaging/core/messaging-controller.ts
