+++ apps/desktop/src/main/messaging/core/messaging-store.ts
