+++ apps/desktop/src/main/messaging/core/model-interaction-mapper.ts
