+++ packages/agent-core/src/index.ts
@@ -53,38 +53,6 @@ export {
   type StoredMessage,
 } from "./persistence/grok-rollout-store.js";
 export {
-  buildMessagingConversationKey,
-  MessagingStore,
-} from "./messaging/messaging-store.js";
-export {
-  CURRENT_MESSAGING_STORE_VERSION,
-  migrateMessagingStoreData,
-  type MessagingDeliveryRecord,
-  type MessagingStoreData,
-} from "./messaging/messaging-migrations.js";
-export {
-  MessagingController,
-  type MessagingControllerOptions,
-} from "./messaging/messaging-controller.js";
-export type {
-  MessagingAdapter,
-  MessagingBackendBridge,
-  MessagingBackendEventListener,
-  MessagingInboundListener,
-} from "./messaging/messaging-adapter.js";
-export {
-  DeterministicInteractionMapper,
-  actionsForIntent,
-} from "./messaging/deterministic-interaction-mapper.js";
-export {
-  ModelInteractionMapper,
-  type ModelInteractionMapperClient,
-} from "./messaging/model-interaction-mapper.js";
-export type {
-  MessagingInteractionMapper,
-  MessagingInteractionMapperResult,
-} from "./messaging/interaction-mapper.js";
-export {
   asObjectArguments,
   readOptionalBoolean,
   readOptionalPositiveInteger,
